@@ -10,8 +10,10 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=10"
+		"node": ">=14.16"
 	},
 	"scripts": {
 		"test": "xo && ava"
@@ -30,15 +32,16 @@
 		"html"
 	],
 	"dependencies": {
-		"marked": "^2.0.1",
+		"easy-transform-stream": "^1.0.0",
+		"marked": "^4.0.16",
 		"plugin-error": "^1.0.1",
 		"through2": "^4.0.2"
 	},
 	"devDependencies": {
-		"ava": "^2.4.0",
-		"p-event": "^4.2.0",
+		"ava": "^4.3.0",
+		"p-event": "^5.0.1",
 		"vinyl": "^2.2.1",
-		"xo": "^0.38.2"
+		"xo": "^0.50.0"
 	},
 	"peerDependencies": {
 		"gulp": ">=4"
